Got our first device nice and showing up (TODO: Add suffixes!).

+++ build_i386.sh
@@ -1,9 +1,17 @@
 #!/usr/bin/sh
 
+set -xe
 source ./prebuild.sh
 
-rm -rf build
-mkdir -p build
+if test "$#" -eq 1 && test "$1" = "--dirty"; then
+  export DIRTY=1
+else
+  read -p "[nop] By pressing Enter, you will be deleting all previous test images and executables. Do you want to proceed?"
+  export DIRTY=0
+  
+  rm -rf build
+  mkdir -p build
+fi
 
 rm -f $(find ./nop -name "*.o") $(find ./lib -name "*.o")
 
@@ -21,23 +29,25 @@ find ./lib -name "*.asm" -exec nasm ${ASMFLAGS}{} -o {}.o \;
 i386-coff-gcc -T files/i386/linker.ld -o build/nop_i386.exe -ffreestanding -O2 -nostdlib -lgcc \
   $(find ./nop -name "*.o") $(find ./lib -name "*.o")
 
-# exit 0
-
-dd if=/dev/zero of=build/test_image.img count=266248 conv=fsync status=progress
+if test ${DIRTY} -eq 0; then
+  dd if=/dev/zero of=build/test_image.img count=266248 conv=fsync status=progress
+fi
 
 export LOOP=$(sudo losetup -f)
 mkdir -p mnt
 
 sudo losetup -P ${LOOP} build/test_image.img
 
-sudo sfdisk ${LOOP} << EOF
+if test ${DIRTY} -eq 0; then
+  sudo sfdisk ${LOOP} << EOF
   label: dos
   unit: sectors
   
   2MiB, ,
 EOF
-
-sudo mkfs.nilfs2 ${LOOP}p1
+  
+  sudo mkfs.nilfs2 ${LOOP}p1
+fi
 
 sudo mount ${LOOP}p1 mnt
 sudo mkdir -p mnt/boot/grub
@@ -45,8 +55,10 @@ sudo mkdir -p mnt/boot/grub
 sudo cp build/nop_i386.exe mnt/boot/nop_i386.exe
 sudo cp files/i386/grub.cfg mnt/boot/grub/grub.cfg
 
-sudo grub-install --target=i386-pc --locales=es --modules="part_msdos nilfs2" \
-  --install-modules="part_apple part_msdos part_gpt nilfs2 multiboot2" --boot-directory=mnt/boot ${LOOP}
+if test ${DIRTY} -eq 0; then
+  sudo grub-install --target=i386-pc --locales=es --modules="part_msdos nilfs2" \
+    --install-modules="part_apple part_msdos part_gpt nilfs2 multiboot2" --boot-directory=mnt/boot ${LOOP}
+fi
 
 sudo umount mnt
